pyproject.toml
[tool.poetry]
name = "db-evolution-system"
version = "1.0.0"
description = "LLM-powered database schema evolution with RAG"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
psycopg2-binary = "^2.9.9"
numpy = "^1.26.0"
faiss-cpu = "^1.7.4"
openai = "^1.3.0"
anthropic = "^0.7.0"
click = "^8.1.7"
fastapi = "^0.104.1"
uvicorn = "^0.24.0"
pydantic = "^2.5.0"
sqlparse = "^0.4.4"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-cov = "^4.1.0"
black = "^23.11.0"
mypy = "^1.7.1"
ruff = "^0.1.6"

[tool.poetry.scripts]
dbevolve = "src.presentation.cli.commands:cli"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
